+++ src/KardFinancial.Test/Unit/MockServer/Organizations/ContentStrategies/ListTest.cs
@@ -21,10 +21,7 @@ public class ListTest : BaseMockServerTest
                   "attributes": {
                     "name": "name",
                     "organizationId": "organizationId",
-                    "filters": [
-                      "NEWLY_LIVE",
-                      "NEWLY_LIVE"
-                    ],
+                    "filter": "NEWLY_LIVE",
                     "categories": [
                       "Arts & Entertainment",
                       "Arts & Entertainment"
@@ -47,10 +44,7 @@ public class ListTest : BaseMockServerTest
                   "attributes": {
                     "name": "name",
                     "organizationId": "organizationId",
-                    "filters": [
-                      "NEWLY_LIVE",
-                      "NEWLY_LIVE"
-                    ],
+                    "filter": "NEWLY_LIVE",
                     "categories": [
                       "Arts & Entertainment",
                       "Arts & Entertainment"