+++ src/KardFinancial.Test/Unit/Serialization/CreateContentStrategyRequestBodyTest.cs
@@ -19,10 +19,7 @@ public class CreateContentStrategyRequestBodyTest
                 "type": "contentStrategy",
                 "attributes": {
                   "name": "Featured Travel",
-                  "filters": [
-                    "HIGHEST_CASHBACK",
-                    "NEWLY_LIVE"
-                  ],
+                  "filter": "HIGHEST_CASHBACK",
                   "categories": [
                     "Travel"
                   ],
@@ -44,11 +41,7 @@ public class CreateContentStrategyRequestBodyTest
                 Attributes = new CreateContentStrategyAttributes
                 {
                     Name = "Featured Travel",
-                    Filters = new List<ContentStrategyFilter>()
-                    {
-                        ContentStrategyFilter.HighestCashback,
-                        ContentStrategyFilter.NewlyLive,
-                    },
+                    Filter = ContentStrategyFilter.HighestCashback,
                     Categories = new List<CategoryOption>() { CategoryOption.Travel },
                     CategoryExclusions = new List<CategoryOption>() { CategoryOption.Gas },
                     MerchantExclusions = new List<string>() { "merchant-abc" },
@@ -68,10 +61,7 @@ public class CreateContentStrategyRequestBodyTest
                 "type": "contentStrategy",
                 "attributes": {
                   "name": "Featured Travel",
-                  "filters": [
-                    "HIGHEST_CASHBACK",
-                    "NEWLY_LIVE"
-                  ],
+                  "filter": "HIGHEST_CASHBACK",
                   "categories": [
                     "Travel"
                   ],