+++ src/KardFinancial/Organizations/ContentStrategies/Types/ContentStrategyAttributes.cs
@@ -28,11 +28,10 @@ public record ContentStrategyAttributes : IJsonOnDeserialized
     public required string OrganizationId { get; set; }
 
     /// <summary>
-    /// Filters applied when selecting offers for the strategy
+    /// Filter applied when selecting offers for the strategy
     /// </summary>
-    [JsonPropertyName("filters")]
-    public IEnumerable<ContentStrategyFilter> Filters { get; set; } =
-        new List<ContentStrategyFilter>();
+    [JsonPropertyName("filter")]
+    public ContentStrategyFilter? Filter { get; set; }
 
     /// <summary>
     /// Merchant categories to include